+++ common/src/main/java/net/mehvahdjukaar/jeed/mixins/DisplayEffectScreenMixin.java
@@ -2,6 +2,7 @@ package net.mehvahdjukaar.jeed.mixins;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.mehvahdjukaar.jeed.Jeed;
+import net.mehvahdjukaar.jeed.compat.IInventoryScreenExtension;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.gui.screens.inventory.EffectRenderingInventoryScreen;
 import net.minecraft.network.chat.Component;
@@ -18,7 +19,7 @@ public abstract class DisplayEffectScreenMixin<T extends AbstractContainerMenu> 
 
     @Override
     public boolean mouseClicked(double x, double y, int activeButton) {
-        if (Jeed.MOD_COMPAT.handleEffectMouseClicked(this, x, y, activeButton)) return true;
+        if (IInventoryScreenExtension.INSTANCE.handleEffectMouseClicked(this, x, y, activeButton)) return true;
 
         return super.mouseClicked(x, y, activeButton);
     }
@@ -26,7 +27,7 @@ public abstract class DisplayEffectScreenMixin<T extends AbstractContainerMenu> 
     @Override
     protected void renderTooltip(PoseStack matrixStack, int x, int y) {
         if (this.hoveredSlot == null && this.menu.getCarried().isEmpty()) {
-            Jeed.MOD_COMPAT.handleEffectRenderTooltip(this, matrixStack, x, y);
+            IInventoryScreenExtension.INSTANCE.handleEffectRenderTooltip(this, matrixStack, x, y);
         }
         super.renderTooltip(matrixStack, x, y);