+++ common/src/main/java/net/mehvahdjukaar/jeed/plugin/rei/ingredient/EffectInstanceDefinition.java
@@ -18,6 +18,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.alchemy.PotionUtils;
@@ -91,7 +92,7 @@ public class EffectInstanceDefinition implements EntryDefinition<MobEffectInstan
 
     @Override
     public boolean equals(MobEffectInstance o1, MobEffectInstance o2, ComparisonContext context) {
-        return o2.equals(o1);
+        return o2.getEffect() == o1.getEffect();
     }
 
     @Override