+++ fabric/src/main/java/net/mehvahdjukaar/jeed/compat/fabric/StylishEffectsScreenExtension.java
@@ -1,30 +1,28 @@
-package net.mehvahdjukaar.jeed.compat.forge;
+package net.mehvahdjukaar.jeed.compat.fabric;
 
 import fuzs.stylisheffects.api.client.MobEffectWidgetContext;
 import fuzs.stylisheffects.api.client.StylishEffectsClientApi;
-import fuzs.stylisheffects.api.client.event.MobEffectWidgetEvent;
-import net.mehvahdjukaar.jeed.compat.IModCompat;
+import fuzs.stylisheffects.api.client.event.MobEffectWidgetEvents;
+import net.mehvahdjukaar.jeed.Jeed;
+import net.mehvahdjukaar.jeed.compat.IInventoryScreenExtension;
 import net.mehvahdjukaar.jeed.plugin.jei.ingredient.EffectInstanceRenderer;
-import net.mehvahdjukaar.jeed.plugin.jei.plugins.InventoryScreenHandler;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.effect.MobEffectInstance;
-import net.minecraftforge.common.MinecraftForge;
 
 import java.util.List;
 
-public class StylishEffectsCompat implements IModCompat {
+public class StylishEffectsScreenExtension implements IInventoryScreenExtension {
 
     @Override
     public void registerHandlers() {
-        MinecraftForge.EVENT_BUS.addListener((final MobEffectWidgetEvent.MouseClicked evt) -> {
-            InventoryScreenHandler.onClickedEffect(evt.getContext().effectInstance(), evt.getMouseX(), evt.getMouseY(), evt.getButton());
-            evt.setCanceled(true);
+        MobEffectWidgetEvents.CLICKED.register((evt, screen, x, y, button) -> {
+            Jeed.PLUGIN.onClickedEffect(evt.effectInstance(), x, y, button);
+            return true;
         });
-        MinecraftForge.EVENT_BUS.addListener((final MobEffectWidgetEvent.EffectTooltip evt) -> {
-            List<Component> lines = evt.getTooltipLines();
+        MobEffectWidgetEvents.TOOLTIP.register((evt, lines, flag) -> {
             lines.clear();
-            List<Component> newTooltip = EffectInstanceRenderer.INSTANCE.getTooltipsWithDescription(evt.getContext().effectInstance(), evt.getTooltipFlag(), false);
+            List<Component> newTooltip = EffectInstanceRenderer.getTooltipsWithDescription(evt.effectInstance(), flag, false);
             lines.addAll(newTooltip);
         });
     }