+++ forge/src/main/java/net/mehvahdjukaar/jeed/forge/JeedImpl.java
@@ -1,24 +1,24 @@
 package net.mehvahdjukaar.jeed.forge;
 
 import net.mehvahdjukaar.jeed.Jeed;
-import net.mehvahdjukaar.jeed.compat.IModCompat;
-import net.mehvahdjukaar.jeed.compat.NativeModCompat;
-import net.mehvahdjukaar.jeed.compat.forge.StylishEffectsCompat;
+import net.mehvahdjukaar.jeed.compat.IInventoryScreenExtension;
+import net.mehvahdjukaar.jeed.compat.forge.StylishEffectsScreenExtension;
+import net.mehvahdjukaar.jeed.compat.forge.VanillaScreenExtension;
 import net.mehvahdjukaar.jeed.recipes.EffectProviderRecipe;
 import net.mehvahdjukaar.jeed.recipes.PotionProviderRecipe;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
-import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraft.world.item.crafting.RecipeType;
-import net.minecraftforge.client.ForgeHooksClient;
+import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.fml.loading.FMLEnvironment;
+import net.minecraftforge.fml.loading.RuntimeDistCleaner;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
@@ -26,7 +26,6 @@ import net.minecraftforge.registries.RegistryObject;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
 
 /**
  * Author: MehVahdJukaar
@@ -45,7 +44,9 @@ public class JeedImpl {
 
         createConfigs();
 
-        Jeed.init();
+        if(FMLEnvironment.dist == Dist.CLIENT) {
+            IInventoryScreenExtension.INSTANCE.registerHandlers();
+        }
     }
 
     private static ForgeConfigSpec.BooleanValue effectBox;
@@ -85,12 +86,12 @@ public class JeedImpl {
     }
 
 
-    public static IModCompat initModCompat() {
+    public static IInventoryScreenExtension initModCompat() {
         //credits to Fuzss for all the Stylish Effects mod compat
         if (ModList.get().isLoaded("stylisheffects")) {
-            return new StylishEffectsCompat();
+            return new StylishEffectsScreenExtension();
         } else {
-            return new NativeModCompat();
+            return new VanillaScreenExtension();
         }
     }
 
@@ -127,43 +128,5 @@ public class JeedImpl {
         return effectColor.get();
     }
 
-    public static MobEffectInstance getHoveredEffect(AbstractContainerScreen<?> screen, double mouseX, double mouseY, boolean ignoreIfSmall) {
-        int minX;
-        boolean cancelShift = false;
-        if (cancelShift)
-            minX = (screen.width - screen.getXSize()) / 2;
-        else
-            minX = screen.getGuiLeft() + screen.getXSize() + 2;
-        int x = screen.width - minX;
-        Collection<MobEffectInstance> collection = Minecraft.getInstance().player.getActiveEffects();
-        if (!collection.isEmpty() && x >= 32) {
-
-            boolean full = x >= 120;
-            var event = ForgeHooksClient.onScreenPotionSize(screen, x, !full, minX);
-            if (event.isCanceled()) return null;
-            full = !event.isCompact();
-            minX = event.getHorizontalOffset();
-            if (!full && ignoreIfSmall) return null;
-            int width = full ? 120 : 32;
-            if (mouseX > minX && mouseX < minX + width) {
-
-                int spacing = 33;
-                if (collection.size() > 5) {
-                    spacing = 132 / (collection.size() - 1);
-                }
-
-
-                List<MobEffectInstance> iterable = collection.stream().filter(ForgeHooksClient::shouldRenderEffect).sorted().collect(Collectors.toList());
-
-                int minY = screen.getGuiTop();
-                int maxHeight = iterable.size() * spacing;
-
-                if (mouseY > minY && mouseY < minY + maxHeight) {
-                    return iterable.get((int) ((mouseY - minY) / spacing));
-                }
-            }
-        }
-        return null;
-    }
 
 }